remove model, likely some more breakage with this, but it still renders
